requirements.txt
streamlit
openai
yfinance
pandas
python-dotenv
openpyxl
plotly